Fix path yet again
It turns out the Ubuntu packages differ from the Debian ones, due to
Debian bug#531592 and a difference of opinion between the Debian
maintainer of PC/SC and the Ubuntu maintainers. As such, if we want to
work on both distributions, we need to add both directories.

Hardcoding paths is silly, Oracle.